+++ packages/cli/templates/modules/src/personal-center/pages/settings/index.tsx
@@ -1,0 +1,5 @@
+import { BasicContent } from "@oj-module/runtime";
+
+export default function Settings() {
+	return <BasicContent>Settings</BasicContent>;
+}